Revert "Modify updateState in vehicle class - dynamic numPicK in solver for anytime"
This reverts commit 6bcd6ee1f0dcb2e0c195f593d245a78d9bd299b0.

+++ DARP_IPS/src/data/Parameters.cpp
@@ -10,7 +10,7 @@
 
 // Constructor and Destructor
 Parameters::Parameters(float alphaParam, float betaParam, float deltaPram, int epochLength, int penaltyL,
-                       int committedTime, int nbThreads, InitialDual initialDual, MainAlgorithm mainAlgorithm,
+                       float committedTime, int nbThreads, InitialDual initialDual, MainAlgorithm mainAlgorithm,
                        int numIter, bool greedyReOptimize, int saveScratch, bool vehicleReturn, float timeWindow,
                        warmStart initialStart, int MIP_maxIncDegree, int CP_IncDegree,
                        bool useMultiStage, float minImp, bool useZoom, int nbColumn,
